Populate `host` and `webURL`fields of FrameworkInfo.
This patch extracts out `FrameworkInfo` construction from the `DriverSettings`
data class to a factory class. This factory class combines the base info
constructed via CLI arguments with the HTTP server's host and port information.
This allows us to populate the `host` and `weburl` fields of framework info,
which enhance the Mesos UI.

This is necessary for users of the `V1_DRIVER` as the new driver does not
automatically populate the `host` field. Further, by using our own host and port
information, we ensure the information in ZooKeeper, the information used for
HTTP redirects and the information in the Mesos UI are all in sync.

Note that in vagrant, the hostname and URL are `aurora.local` because we set the
`--hostname` argument of the scheduler. By default Java will set it to the FQDN
or IP address of the host.

Testing Done:
e2e tests.

Bugs closed: AURORA-1905

Reviewed at https://reviews.apache.org/r/57708/