Fix schema to allow multiple task volumes per task.
The original commit adding this feature added an artifical constraint to the
schema that prevented more than one task volume per task. This is because there
was a `UNIQUE` constraint between the volumes table and the task config table,
preventing a task config from being associated with more than one volume.

This patch removes that constraint. As a result some of the MyBatis mappers had
to change and a new migration was added.

Bugs closed: AURORA-1914

Reviewed at https://reviews.apache.org/r/58066/