Bump initial_task_kill_retry_interval to 15s.
It is not very common that kills are dropped by Mesos and have to be retried
by Aurora. It therefore makes sense to slightly increase the retry timeout
so that we don't retry needlessly when Thermos is still busy executing
the lifecycle methods.

By default, Thermos uses the following kill escalation sequence:

  * /quitquitquit
  * wait 5s
  * /abortabortabort
  * wait 5s
  * SIGTERM
  * wait up to 1 minute
  * SIGKILL

Reviewed at https://reviews.apache.org/r/58611/