Fix concurrency issues around banned offers in HostOffers
Currently, `getOffers` and `getWeaklyConsistentOffers` provide iterables that
use lambdas (in filter) not backed by concurrent data structures
(`globallyBannedOffers` and `staticallyBannedOffers`). This can lead to a race
condition where we have unknown behavior if reading and modification happens at
the same time.

To fix this, in `getOffers` we will revert to the previous behavior (before
patch https://reviews.apache.org/r/61804/) where we create a copy of the current
offers as well as a copy of the banned offers for consistency.

For `getWeaklyConsistentOffers`, we will use concurrent data structures for
`globallyBannedOffers` and `staticallyBannedOffers` so we don't have a race that
may produce unknown side effects.

Reviewed at https://reviews.apache.org/r/61918/