Run Jenkins tests without the Gradle daemon
This follows the recommendation of Gradle to only use their daemon when
running in local environments but not in CI environments.

We are seeing spurious build failures from time to time on the shared
Apache build server. Disabling the daemon might help to prevent those.

https://docs.gradle.org/current/userguide/gradle_daemon.html#when_should_i_not_use_the_gradle_daemon

Reviewed at https://reviews.apache.org/r/62832/