Fix flaky WebhookTest
I opted for a minimal patch that stays close to the style of the other
tests. There is a remaining risk of a race condition if the OS re-uses
the same port between stopping jetty and triggering the webhooks, but
it should be very unlikely.

Reviewed at https://reviews.apache.org/r/64482/