Switch Thermos to lazy log formatting
This is the first part of a small series of Thermos observer performance
improvements.

As a first iteration, this switches all logging to use the logger-embedded
formatting rather than doing it eager up front. This has the advantage that
we produce less garbage if debug logging is disabled.

Reviewed at https://reviews.apache.org/r/66136/