Prune updates that have no surviving job keys in the TaskStore.
We are running into a situation where we have a lot of short-lived ad-hoc services launched and their updates are sticking around for 30 days, even though the tasks are garbage collected much sooner. This change picks up those updates and prunes them as soon as the tasks are gone.

Reviewed at https://reviews.apache.org/r/68071/