Memoize loader.load and loader.load_json (#41)
* Memoize loader.load and loader.load_json

Problem: when reusing the aurora.client.cli.context to process multiple jobkeys from a config_file, the client currently reloads and processes the config file for each jobkey. For large complicated aurora files this can take ~40s to process which makes it expensive to do this every time for each jobkey.

Solution: memoize the loader.load() on the config file path so that we only load and process each config file once.

Result: We only need to load and process each config file once. For config files with ~50keys and a 40s load time, this significantly reduces the overall time spent inspecting the jobs.