Auto Pause for Batch based update strategies (#54)
Auto pause feature for VariableBatch strategy and Batch strategy. With this feature enabled, when an update is ROLLING_FORWARD, the update will automatically pause itself right before a new batch is started. 

(This feature is being released as tested but in beta state. We are looking to collect feedback before we consider it fully stable.)